render: fix undefined behavior in shift operations
Add explicit unsigned casts to left shift operations to avoid signed
integer overflow undefined behavior. This addresses UBSan warnings
about shifts that exceed the width of signed int type.

Also fix TIME_HALF constant in event.c to use unsigned literal.

+++ obrender/color.c
@@ -66,16 +66,16 @@ RrColor* RrColorNew(const RrInstance* inst, gint r, gint g, gint b) {
   g_assert(g >= 0 && g < 256);
   g_assert(b >= 0 && b < 256);
 
-  key = (r << 24) + (g << 16) + (b << 8);
+  key = (gint)(((guint)r << 24) + ((guint)g << 16) + ((guint)b << 8));
 #ifndef NO_COLOR_CACHE
   if ((out = g_hash_table_lookup(RrColorHash(inst), &key))) {
     out->refcount++;
   }
   else {
 #endif
-    xcol.red = (r << 8) | r;
-    xcol.green = (g << 8) | g;
-    xcol.blue = (b << 8) | b;
+    xcol.red = ((guint)r << 8) | r;
+    xcol.green = ((guint)g << 8) | g;
+    xcol.blue = ((guint)b << 8) | b;
     if (XAllocColor(RrDisplay(inst), RrColormap(inst), &xcol)) {
       out = g_slice_new(RrColor);
       out->inst = inst;
@@ -137,7 +137,7 @@ void RrReduceDepth(const RrInstance* inst, RrPixel32* data, XImage* im) {
             r = (data[x] >> RrDefaultRedOffset) & 0xFF;
             g = (data[x] >> RrDefaultGreenOffset) & 0xFF;
             b = (data[x] >> RrDefaultBlueOffset) & 0xFF;
-            p32[x] = (r << ro) + (g << go) + (b << bo);
+            p32[x] = ((guint)r << ro) + ((guint)g << go) + ((guint)b << bo);
           }
           data += im->width;
           p32 += im->width;
@@ -176,7 +176,7 @@ void RrReduceDepth(const RrInstance* inst, RrPixel32* data, XImage* im) {
           g = g >> gs;
           b = (data[x] >> RrDefaultBlueOffset) & 0xFF;
           b = b >> bs;
-          p16[x] = (r << ro) + (g << go) + (b << bo);
+          p16[x] = ((guint)r << ro) + ((guint)g << go) + ((guint)b << bo);
         }
         data += im->width;
         p16 += im->bytes_per_line / 2;
@@ -192,7 +192,7 @@ void RrReduceDepth(const RrInstance* inst, RrPixel32* data, XImage* im) {
             g = g >> gs;
             b = (data[x] >> RrDefaultBlueOffset) & 0xFF;
             b = b >> bs;
-            p8[x] = (r << ro) + (g << go) + (b << bo);
+            p8[x] = ((guint)r << ro) + ((guint)g << go) + ((guint)b << bo);
           }
           data += im->width;
           p8 += im->bytes_per_line;
@@ -219,7 +219,8 @@ XColor* RrPickColor(const RrInstance* inst, gint r, gint g, gint b) {
   r = (r & 0xff) >> (8 - RrPseudoBPC(inst));
   g = (g & 0xff) >> (8 - RrPseudoBPC(inst));
   b = (b & 0xff) >> (8 - RrPseudoBPC(inst));
-  return &RrPseudoColors(inst)[(r << (2 * RrPseudoBPC(inst))) + (g << (1 * RrPseudoBPC(inst))) + b];
+  return &RrPseudoColors(
+      inst)[((guint)r << (2 * RrPseudoBPC(inst))) + ((guint)g << (1 * RrPseudoBPC(inst))) + (guint)b];
 }
 
 static void swap_byte_order(XImage* im) {
@@ -275,8 +276,8 @@ void RrIncreaseDepth(const RrInstance* inst, RrPixel32* data, XImage* im) {
           r = (p32[x] >> RrRedOffset(inst)) & 0xff;
           g = (p32[x] >> RrGreenOffset(inst)) & 0xff;
           b = (p32[x] >> RrBlueOffset(inst)) & 0xff;
-          data[x] = (r << RrDefaultRedOffset) + (g << RrDefaultGreenOffset) + (b << RrDefaultBlueOffset) +
-                    (0xff << RrDefaultAlphaOffset);
+          data[x] = ((guint)r << RrDefaultRedOffset) + ((guint)g << RrDefaultGreenOffset) +
+                    ((guint)b << RrDefaultBlueOffset) + ((guint)0xff << RrDefaultAlphaOffset);
         }
         data += im->width;
         p32 += im->bytes_per_line / 4;
@@ -285,11 +286,11 @@ void RrIncreaseDepth(const RrInstance* inst, RrPixel32* data, XImage* im) {
     case 16:
       for (y = 0; y < im->height; y++) {
         for (x = 0; x < im->width; x++) {
-          r = (p16[x] & RrRedMask(inst)) >> RrRedOffset(inst) << RrRedShift(inst);
-          g = (p16[x] & RrGreenMask(inst)) >> RrGreenOffset(inst) << RrGreenShift(inst);
-          b = (p16[x] & RrBlueMask(inst)) >> RrBlueOffset(inst) << RrBlueShift(inst);
-          data[x] = (r << RrDefaultRedOffset) + (g << RrDefaultGreenOffset) + (b << RrDefaultBlueOffset) +
-                    (0xff << RrDefaultAlphaOffset);
+          r = (gint)(((guint)(p16[x] & RrRedMask(inst)) >> RrRedOffset(inst)) << RrRedShift(inst));
+          g = (gint)(((guint)(p16[x] & RrGreenMask(inst)) >> RrGreenOffset(inst)) << RrGreenShift(inst));
+          b = (gint)(((guint)(p16[x] & RrBlueMask(inst)) >> RrBlueOffset(inst)) << RrBlueShift(inst));
+          data[x] = ((guint)r << RrDefaultRedOffset) + ((guint)g << RrDefaultGreenOffset) +
+                    ((guint)b << RrDefaultBlueOffset) + ((guint)0xff << RrDefaultAlphaOffset);
         }
         data += im->width;
         p16 += im->bytes_per_line / 2;
@@ -302,7 +303,7 @@ void RrIncreaseDepth(const RrInstance* inst, RrPixel32* data, XImage* im) {
       for (y = 0; y < im->height; y++) {
         for (x = 0; x < im->width; x++) {
           if (!(((p8[x / 8]) >> (x % 8)) & 0x1))
-            data[x] = 0xff << RrDefaultAlphaOffset; /* black */
+            data[x] = (guint)0xff << RrDefaultAlphaOffset; /* black */
           else
             data[x] = 0xffffffff; /* white */
         }
