+++ openbox/focus_cycle_popup.c
@@ -166,8 +166,8 @@ void focus_cycle_popup_startup(gboolean reconfig) {
     gint x, y, o;
 
     tc = ob_rr_theme->osd_text_active_color;
-    color = ((tc->r & 0xff) << RrDefaultRedOffset) + ((tc->g & 0xff) << RrDefaultGreenOffset) +
-            ((tc->b & 0xff) << RrDefaultBlueOffset);
+    color = ((guint)(tc->r & 0xff) << RrDefaultRedOffset) + ((guint)(tc->g & 0xff) << RrDefaultGreenOffset) +
+            ((guint)(tc->b & 0xff) << RrDefaultBlueOffset);
 
     o = 0;
     for (x = 0; x < HILITE_SIZE; x++)
@@ -184,7 +184,7 @@ void focus_cycle_popup_startup(gboolean reconfig) {
           a = 0x22;
         }
 
-        p[o++] = color + (a << RrDefaultAlphaOffset);
+        p[o++] = color + ((guint)a << RrDefaultAlphaOffset);
       }
   }
